* JwtTokenHandler no longer throws an exception if Java doesn't like the token sent * Moved Addresses to its own Controller/DAO/JDBC * Email body can now be read from HTML files * Renamed /api/event/{eventid}/sendinvite to /api/event/{eventid}/invite * API calls now override the object with path and userid variables * Added emailExists method in the DAO/JDBC * Event POJO validation no longer errors out with NullPointerException * Changed Invitee POJO's eventId to a long

+++ vue_starter_java/backend/src/main/java/com/techelevator/model/pojo/Invitee.java
@@ -9,7 +9,7 @@ public class Invitee {
 	@NotNull(message="Email must be provided")
 	private String email;
 	
-	private int eventId;
+	private long eventId;
 	
 	@NotNull(message="Role must be assigned")
 	private String role;
@@ -26,10 +26,10 @@ public class Invitee {
 	public void setEmail(String email) {
 		this.email = email;
 	}
-	public int getEventId() {
+	public long getEventId() {
 		return eventId;
 	}
-	public void setEventId(int eventId) {
+	public void setEventId(long eventId) {
 		this.eventId = eventId;
 	}
 	public String getRole() {
